+++ src/components/Pages/ArenaPage.tsx
@@ -1,0 +1,3 @@
+export const ArenaPage = () => {
+  return <div>FIGHT!</div>;
+};