+++ src/styling/BaphThemeProvider.tsx
@@ -1,0 +1,18 @@
+import { ThemeProvider as EmotionThemeProvider, Theme } from '@emotion/react';
+import { ReactNode } from 'react';
+
+type ThemeProviderTypes = {
+  theme?: Theme;
+  children?: ReactNode;
+};
+const BaphThemeProvider: React.FC<ThemeProviderTypes> = ({
+  theme,
+  children
+}) => {
+  console.log;
+  return (
+    <EmotionThemeProvider theme={theme || {}}>{children}</EmotionThemeProvider>
+  );
+};
+
+export default BaphThemeProvider;